+++ src/common/Dataset.h
@@ -27,8 +27,8 @@ namespace platform {
         std::pair<vector<std::vector<float>>&, std::vector<int>&> getVectors();
         std::pair<torch::Tensor&, torch::Tensor&> getTensors();
         std::tuple<torch::Tensor&, torch::Tensor&, torch::Tensor&, torch::Tensor&> getTrainTestTensors(std::vector<int>& train, std::vector<int>& test);
-        int getNFeatures() const;
-        int getNSamples() const;
+        long getNFeatures() const;
+        long getNSamples() const;
         std::vector<bool>& getNumericFeatures() { return numericFeatures; }
         void load();
         const bool inline isLoaded() const { return loaded; };
@@ -37,7 +37,7 @@ namespace platform {
         std::string name;
         fileType_t fileType;
         std::string className;
-        int n_samples{ 0 }, n_features{ 0 };
+        long n_samples{ 0 }, n_features{ 0 };
         std::vector<int> numericFeaturesIdx;
         std::string discretizer_algorithm;
         std::vector<bool> numericFeatures; // true if feature is numeric
